Add files via upload
The file Kuferefeta-v3.docm is a macro enabled word document. It contains code that runs a very simplistic spell checker for Shona. The approach is partly inspired by the work of Peter Norvig who described how to write a spell checker in Python. I have used Word and written the spell checker as a macro because Microsoft word is widely distributed among the users of Shona and because the spell checking innards of Word are closed source. Most of the buttons that come up when the "dialogue box"/form with the word corrections is presented are still work in progress, but the key part of the spelling corrector now works. The key difference between my approach and that which Peter Norvig proposes is that I am not displaying the most likely correct word, instead, I present a list of all possible correct words. There is also a bug in the search algorithm which is leading to more false positives. For the "bag of words", I downloaded a publicly available Corpus of Shona words and pre-processed them into a list that I can reference. They are in the file mazwi_akawururwa.txt. A better corpus could potentially improve the results, and so would a better search algorithm. These will be improved in later versions.